refactor: break every runtime import cycle
madge (type-only and dynamic imports skipped) reported 13 cycles; now 0.

- providers <-> stores: providerRuntime no longer imports dataStore/query.
  It exposes onCodexSessionChange() and main.ts installs the listener
  that invalidates the affected providers' queries. providers/openai reads
  configured models through the plugin handle (state.svelte) instead of
  the dataStore module, and state.svelte itself is now a leaf: the
  thinkingProcessPref accessor that needed getData() moved to its own
  module. The unused ProviderRegistry re-export that closed
  providers/index <-> providers/registry is gone.
- dataStore <-> utils/{agentPaths,fileFiltering}: the utils read the agent
  folder and agent names from an injected AgentPathSource
  (utils/agentPathSource.ts) that PluginDataStore installs on
  construction, instead of importing the store that imports them.
- lib/pendingChangeNavigation <-> editor/readingViewDiffProcessor: the pure
  line-math helpers both sides share moved to editor/diffLineMath.ts.

Tests that mocked getData() to feed the path utils now install a stand-in
AgentPathSource.
