+++ test/agent/builtInToolDefaults.test.ts
